Complete voting feature for games
Change a design choice to display votes, make an adjustment in the backend to store double values

+++ tffconnect_backend/tffconnect/games/migrations/0003_games_rating_count_games_referee_rating.py
@@ -18,6 +18,6 @@ class Migration(migrations.Migration):
         migrations.AddField(
             model_name='games',
             name='referee_rating',
-            field=models.IntegerField(default=0),
+            field=models.FloatField(default=0),
         ),
     ]
