+++ tffconnect_backend/tffconnect/games/models.py
@@ -7,7 +7,7 @@ class Games(models.Model):
     game_name = models.CharField(max_length=100)
     game_date = models.DateField('date played')
     game_result = models.CharField(max_length=100)
-    referee_rating = models.IntegerField(default=0)
+    referee_rating = models.FloatField(default=0)
     rating_count = models.IntegerField(default=0)
 
     def __str__(self):
